refactor: extract FullAnalysisRunner + frontier-queue scheduler
DiagramGenerator.generate_analysis (41 lines) and _generate_subcomponents
(78 lines) moved into a new FullAnalysisRunner class
(diagram_analysis/full_analysis_runner.py). The frontier-queue scheduler
that submits children as soon as parents complete is the algorithmic
core of full analysis; it now lives next to its only caller.

DiagramGenerator.generate_analysis becomes a thin delegate. The class
now holds shared state (agents, static analysis, file coverage) and
exposes helpers (process_component, _build_file_coverage_summary,
_normalize_repo_path) that both runners reach into via the
collaborator-style constructor. process_component stays here because
codeboarding_workflows.analysis.run_partial calls it externally.

Five imports (FIRST_COMPLETED, Future, wait, nullcontext, os) are
unused after the move — dropped.

Combined with the prior IncrementalRunner extraction:
- DiagramGenerator: 524 -> 370 lines
- generate_analysis_incremental: ~75 lines extracted -> IncrementalRunner
- generate_analysis + _generate_subcomponents: ~120 lines extracted -> FullAnalysisRunner

Tests: 1557 pass (same 2 pre-existing failures unrelated).
